refactor: add infra package and migration
