+++ api-server/src/main/java/nexters/payout/apiserver/stock/application/StockQueryService.java
@@ -13,7 +13,7 @@ import nexters.payout.domain.dividend.domain.repository.DividendRepository;
 import nexters.payout.domain.stock.domain.Sector;
 import nexters.payout.domain.stock.domain.Stock;
 import nexters.payout.domain.stock.domain.repository.StockRepository;
-import nexters.payout.domain.stock.domain.repository.StockRepositoryCustom;
+import nexters.payout.domain.stock.infra.StockRepositoryCustom;
 import nexters.payout.domain.stock.domain.service.DividendAnalysisService;
 import nexters.payout.domain.stock.domain.service.SectorAnalysisService;
 import nexters.payout.domain.stock.domain.service.SectorAnalysisService.SectorInfo;
@@ -32,13 +32,12 @@ import java.util.stream.Collectors;
 public class StockQueryService {
 
     private final StockRepository stockRepository;
-    private final StockRepositoryCustom stockRepositoryCustom;
     private final DividendRepository dividendRepository;
     private final SectorAnalysisService sectorAnalysisService;
     private final DividendAnalysisService dividendAnalysisService;
 
     public List<StockResponse> searchStock(final String keyword, final Integer pageNumber, final Integer pageSize) {
-        return stockRepositoryCustom.findStocksByTickerOrNameWithPriority(keyword, pageNumber, pageSize)
+        return stockRepository.findStocksByTickerOrNameWithPriority(keyword, pageNumber, pageSize)
                 .stream()
                 .map(StockResponse::from)
                 .collect(Collectors.toList());
