+++ api-server/src/test/java/nexters/payout/apiserver/stock/application/StockQueryServiceTest.java
@@ -12,7 +12,7 @@ import nexters.payout.domain.dividend.domain.repository.DividendRepository;
 import nexters.payout.domain.stock.domain.Sector;
 import nexters.payout.domain.stock.domain.Stock;
 import nexters.payout.domain.stock.domain.repository.StockRepository;
-import nexters.payout.domain.stock.domain.repository.StockRepositoryCustom;
+import nexters.payout.domain.stock.infra.StockRepositoryCustom;
 import nexters.payout.domain.stock.domain.service.DividendAnalysisService;
 import nexters.payout.domain.stock.domain.service.SectorAnalysisService;
 import org.junit.jupiter.api.Test;
@@ -45,8 +45,6 @@ class StockQueryServiceTest {
     @Mock
     private StockRepository stockRepository;
     @Mock
-    private StockRepositoryCustom stockRepositoryCustom;
-    @Mock
     private DividendRepository dividendRepository;
     @Spy
     private SectorAnalysisService sectorAnalysisService;
@@ -56,7 +54,7 @@ class StockQueryServiceTest {
     @Test
     void 검색된_종목_정보를_정상적으로_반환한다() {
         // given
-        given(stockRepositoryCustom.findStocksByTickerOrNameWithPriority(any(), any(), any())).willReturn(List.of(StockFixture.createStock(AAPL, Sector.TECHNOLOGY)));
+        given(stockRepository.findStocksByTickerOrNameWithPriority(any(), any(), any())).willReturn(List.of(StockFixture.createStock(AAPL, Sector.TECHNOLOGY)));
 
         // when
         List<StockResponse> actual = stockQueryService.searchStock("A", 1 , 2);
