+++ domain/src/main/java/nexters/payout/domain/stock/infra/StockRepositoryCustom.java
@@ -1,11 +1,10 @@
-package nexters.payout.domain.stock.domain.repository;
+package nexters.payout.domain.stock.infra;
 
 import nexters.payout.domain.stock.domain.Stock;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-@Repository
 public interface StockRepositoryCustom {
 
     List<Stock> findStocksByTickerOrNameWithPriority(String search, Integer pageNumber, Integer pageSize);
